# docs/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerForge Pro API",
    "description": "CV-first career acceleration for South African remote work. Human-supervised: sensitive actions require explicit consent.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Search Jobs",
        "operationId": "search_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(all|wwr|remoteok|remotive|adzuna|user_url)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "sa_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Sa Only"
            }
          },
          {
            "name": "max_age_days",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 90,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Age Days"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(newest|oldest)$",
              "default": "newest",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobOut"
                  },
                  "title": "Response Search Jobs Api V1 Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/to-application": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Job To Application",
        "description": "The hand-off: job -> JD record -> application package (tailored CV +\ncover letter). Consent-gated; nothing is submitted anywhere.",
        "operationId": "job_to_application_api_v1_jobs__job_id__to_application_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Job To Application Api V1 Jobs  Job Id  To Application Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/add-url": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Add Job Url",
        "description": "User-directed: fetch one public job page the user provides.",
        "operationId": "add_job_url_api_v1_jobs_add_url_post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUrlIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/profiles/{profile_id}/saved-searches": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Saved",
        "operationId": "list_saved_api_v1_jobs_profiles__profile_id__saved_searches_get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedSearchOut"
                  },
                  "title": "Response List Saved Api V1 Jobs Profiles  Profile Id  Saved Searches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Save Search",
        "operationId": "save_search_api_v1_jobs_profiles__profile_id__saved_searches_post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/profiles/{profile_id}/saved-searches/{search_id}": {
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Delete Search",
        "operationId": "delete_search_api_v1_jobs_profiles__profile_id__saved_searches__search_id__delete",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          },
          {
            "name": "search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Search Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddUrlIn": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 500,
            "minLength": 10,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "AddUrlIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "tags": {
            "type": "string",
            "title": "Tags"
          },
          "salary_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Text"
          },
          "posted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "open_to_sa": {
            "type": "string",
            "title": "Open To Sa"
          },
          "sa_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sa Signals"
          },
          "global_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Global Signals"
          },
          "exclude_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclude Signals"
          },
          "payment_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Payment Signals"
          },
          "timezone_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Timezone Signals"
          },
          "remote_type": {
            "type": "string",
            "title": "Remote Type"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobMatchOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "title",
          "company",
          "location",
          "url",
          "tags",
          "salary_text",
          "posted_at",
          "fetched_at",
          "open_to_sa",
          "sa_signals",
          "global_signals",
          "exclude_signals",
          "payment_signals",
          "timezone_signals",
          "remote_type",
          "description"
        ],
        "title": "JobDetailOut"
      },
      "JobMatchOut": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "components": {
            "additionalProperties": true,
            "type": "object",
            "title": "Components"
          },
          "skill_hits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skill Hits"
          },
          "keyword_hits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keyword Hits"
          },
          "weights": {
            "additionalProperties": true,
            "type": "object",
            "title": "Weights"
          }
        },
        "type": "object",
        "required": [
          "score",
          "components",
          "skill_hits",
          "keyword_hits",
          "weights"
        ],
        "title": "JobMatchOut"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "tags": {
            "type": "string",
            "title": "Tags"
          },
          "salary_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Text"
          },
          "posted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "open_to_sa": {
            "type": "string",
            "title": "Open To Sa"
          },
          "sa_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sa Signals"
          },
          "global_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Global Signals"
          },
          "exclude_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclude Signals"
          },
          "payment_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Payment Signals"
          },
          "timezone_signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Timezone Signals"
          },
          "remote_type": {
            "type": "string",
            "title": "Remote Type"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobMatchOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "title",
          "company",
          "location",
          "url",
          "tags",
          "salary_text",
          "posted_at",
          "fetched_at",
          "open_to_sa",
          "sa_signals",
          "global_signals",
          "exclude_signals",
          "payment_signals",
          "timezone_signals",
          "remote_type"
        ],
        "title": "JobOut"
      },
...
      "SavedSearchIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name"
          },
          "filters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SavedSearchIn"
      },
      "SavedSearchOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "filters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "filters",
          "created_at"
        ],
        "title": "SavedSearchOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}